refactor(oauth): add neutral onyx/oauth leaf for shared OAuth primitives
Relocate, with no behavior change, the OAuth refresh error taxonomy +
token-response classifier, the clockless expiry helpers, and the
authorization-code-grant wire layer (OAuthFlowParams, the SSRF endpoint guard,
authorize-URL builder, code/refresh-token exchanges) into a new neutral
onyx/oauth/ package.

Previously these lived in onyx/external_apps/providers/base.py,
onyx/external_apps/token_utils.py, and onyx/auth/oauth_token_manager.py, which
forced auth/ and the MCP feature to import *up* into the external_apps feature
package to reuse them. onyx/oauth/ depends only on onyx.utils + onyx.server.security,
so every OAuth consumer imports down into it instead.

Repoints all call sites and tests; deletes the relocated token_utils.py.
