+++ backend/tests/external_dependency_unit/tools/test_oauth_token_manager.py
@@ -18,13 +18,13 @@ from requests import HTTPError
 from requests import Response
 from sqlalchemy.orm import Session
 
-from onyx.auth.oauth_token_manager import build_oauth_authorization_url
-from onyx.auth.oauth_token_manager import exchange_oauth_code_for_token
-from onyx.auth.oauth_token_manager import OAuthFlowParams
 from onyx.auth.oauth_token_manager import OAuthTokenManager
 from onyx.db.models import OAuthConfig
 from onyx.db.oauth_config import create_oauth_config
 from onyx.db.oauth_config import upsert_user_oauth_token
+from onyx.oauth.exchange import build_oauth_authorization_url
+from onyx.oauth.exchange import exchange_oauth_code_for_token
+from onyx.oauth.exchange import OAuthFlowParams
 from onyx.utils.sensitive import SensitiveValue
 from tests.external_dependency_unit.conftest import create_test_user
 
@@ -93,7 +93,7 @@ class TestOAuthTokenManagerValidation:
 
         assert access_token == "token_without_expiry"
 
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_get_valid_access_token_with_expired_token_refreshes(
         self, mock_post: Mock, db_session: Session
     ) -> None:
@@ -153,7 +153,7 @@ class TestOAuthTokenManagerValidation:
 
         assert access_token is None
 
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_get_valid_access_token_refresh_fails(
         self, mock_post: Mock, db_session: Session
     ) -> None:
@@ -182,7 +182,7 @@ class TestOAuthTokenManagerValidation:
 class TestOAuthTokenManagerRefresh:
     """Tests for token refresh functionality"""
 
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_refresh_token_success(self, mock_post: Mock, db_session: Session) -> None:
         """Test successful token refresh"""
         oauth_config = _create_test_oauth_config(db_session)
@@ -223,7 +223,7 @@ class TestOAuthTokenManagerRefresh:
         assert token_data["refresh_token"] == "new_refresh"
         assert "expires_at" in token_data
 
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_refresh_token_preserves_refresh_token(
         self, mock_post: Mock, db_session: Session
     ) -> None:
@@ -261,7 +261,7 @@ class TestOAuthTokenManagerRefresh:
         token_data = user_token.token_data.get_value(apply_mask=False)
         assert token_data["refresh_token"] == "old_refresh"
 
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_refresh_token_http_error(
         self, mock_post: Mock, db_session: Session
     ) -> None:
@@ -344,7 +344,7 @@ class TestOAuthTokenManagerExpiration:
 class TestOAuthTokenManagerCodeExchange:
     """Tests for authorization code exchange"""
 
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_exchange_code_for_token_success(
         self, mock_post: Mock, db_session: Session
     ) -> None:
@@ -389,7 +389,7 @@ class TestOAuthTokenManagerCodeExchange:
         ] == oauth_config.client_secret.get_value(apply_mask=False)
         assert call_args[1]["data"]["redirect_uri"] == "https://example.com/callback"
 
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_exchange_code_for_token_http_error(
         self, mock_post: Mock, db_session: Session
     ) -> None:
@@ -566,8 +566,8 @@ class TestSharedOAuthPrimitives:
         assert "code_challenge_method" not in query
         assert "resource" not in query
 
-    @patch("onyx.auth.oauth_token_manager.validate_oauth_endpoint_url")
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.validate_oauth_endpoint_url")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_exchange_code_sends_code_verifier(
         self, mock_post: Mock, mock_validate: Mock
     ) -> None:
@@ -597,8 +597,8 @@ class TestSharedOAuthPrimitives:
         assert sent["code_verifier"] == "verifier_123"
         assert sent["client_secret"] == "known_client_secret"
 
-    @patch("onyx.auth.oauth_token_manager.validate_oauth_endpoint_url")
-    @patch("onyx.auth.oauth_token_manager.requests.post")
+    @patch("onyx.oauth.exchange.validate_oauth_endpoint_url")
+    @patch("onyx.oauth.exchange.requests.post")
     def test_exchange_code_raises_on_http_error(
         self, mock_post: Mock, mock_validate: Mock
     ) -> None:
