+++ backend/tests/unit/external_apps/test_token_refresh.py
@@ -10,9 +10,9 @@ from redis.exceptions import ConnectionError as RedisConnectionError
 from sqlalchemy.exc import SQLAlchemyError
 
 from onyx.external_apps import token_refresh as tr
-from onyx.external_apps.providers.base import TokenRefreshTerminalError
-from onyx.external_apps.providers.base import TokenRefreshTransientError
 from onyx.external_apps.providers.google_calendar import GoogleCalendarProvider
+from onyx.oauth.errors import TokenRefreshTerminalError
+from onyx.oauth.errors import TokenRefreshTransientError
 
 # ---------------------------------------------------------------------------
 # Provider.refresh_credentials (RFC-6749 default on OAuthExternalAppProvider)
