+++ backend/tests/unit/external_apps/test_token_utils.py
@@ -2,8 +2,8 @@ from datetime import datetime
 from datetime import timedelta
 from datetime import timezone
 
-from onyx.external_apps.token_utils import needs_refresh
-from onyx.external_apps.token_utils import stamp_expires_at
+from onyx.oauth.expiry import needs_refresh
+from onyx.oauth.expiry import stamp_expires_at
 
 _NOW = datetime(2026, 5, 29, 12, 0, 0, tzinfo=timezone.utc)
